Use process.execPath instead of process.argv0 to find node location
The problem of using argv0 is that it depends on the current working
directory while installing the native application which might be
different from the working directory while running. Typically, using
argv0 will break if the node binary is given with a relative
path (e.g., ./node_modules/.bin/node). Instead process.execPath is an
absolute filename.